requirements.txt
numpy==1.24.2
opencv-python==4.6.0.66
packaging==23.0
pdf2image==1.16.3
Pillow==9.5.0
pytesseract==0.3.10
tqdm==4.65.0
editdistance~=0.6.2
evaluate==0.4.0
jiwer==3.0.2
google-cloud-vision==3.4.4
azure-cognitiveservices-vision-computervision==0.9.0
pillow==9.5.0
boto3==1.28.20
beautifulsoup4~=4.12.2
scipy~=1.11.1
matplotlib~=3.7.2
pip~=21.3.1
wheel~=0.37.1
lxml~=4.9.3
setuptools~=60.2.0
msrest~=0.7.1
Jinja2~=3.1.2
pandas~=2.0.3
scikit-learn~=1.3.2
